+++ bidworks_backend/settings.py
@@ -10,7 +10,15 @@ For the full list of settings and their values, see
 https://docs.djangoproject.com/en/5.1/ref/settings/
 """
 
+
+from dotenv import load_dotenv
+import os
 from pathlib import Path
+from datetime import timedelta
+
+#* Carrega os bagui
+load_dotenv()
+
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
@@ -23,28 +31,30 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 SECRET_KEY = 'django-insecure-5cpxv-7&zpmjp26jxhduaa@-g*b#k0io1(xr_h4m%c#mwd%toc'
 
 # SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = True
+DEBUG = False
 
-ALLOWED_HOSTS = []
+ALLOWED_HOSTS = [
+    "127.0.0.1",
+    "bidworks-backend.vercel.app"
+]
 
-
+CORS_ALLOWED_ORIGINS = [
+    "127.0.0.1",
+    "bidworks-backend.vercel.app"
+]
 
 # Application definition
 INSTALLED_APPS = [
     'api',
-    'django.contrib.auth',
-    'django.contrib.contenttypes',
-    'django.contrib.sessions',
 ]
 
+# Middlewares do backend
 MIDDLEWARE = [
-    'django.middleware.security.SecurityMiddleware',
-    'django.contrib.sessions.middleware.SessionMiddleware',
+    'bidworks_backend.middleware.token_middleware.TokenAuthMiddleware',
     'django.middleware.common.CommonMiddleware',
-    'django.middleware.csrf.CsrfViewMiddleware',
-    'django.contrib.auth.middleware.AuthenticationMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
+
 
 ROOT_URLCONF = 'bidworks_backend.urls'
 
@@ -68,15 +78,23 @@ WSGI_APPLICATION = 'bidworks_backend.wsgi.application'
 
 
 # Database
-# https://docs.djangoproject.com/en/5.1/ref/settings/#databases
+# Definido como padrão PostgreSQL do supabase -> se quiser usar o SQlite em desenvolvimento, descomente a linha abaixo
 
 DATABASES = {
     'default': {
-        'ENGINE': 'django.db.backends.sqlite3',
-        'NAME': BASE_DIR / 'db.sqlite3',
+        'ENGINE': 'django.db.backends.postgresql_psycopg2',
+        'NAME': 'postgres',
+        'USER': os.getenv('POSTGRES_TRANSACTION_USER'),
+        'PASSWORD': os.getenv('POSTGRES_PASSWORD'),
+        'HOST': os.getenv('POSTGRES_TRANSACTION_HOST'),
+        'PORT': '6543',
+        'POOL_MODE': 'transaction'
     }
+    # 'default': {
+    #     'ENGINE': 'django.db.backends.sqlite3',
+    #     'NAME': BASE_DIR / 'db.sqlite3',
+    # }
 }
-
 
 # Password validation
 # https://docs.djangoproject.com/en/5.1/ref/settings/#auth-password-validators