+++ changelog.md
@@ -25,3 +25,10 @@
 - Cria um novo usuário -> salva o hash da senha na tabela de credenciais
 - Autentica um usuário via email e senha
 - Obtem um usuario por ID
+
+
+## [0.0.3] - 01/04/2025
+
+- Remoção de serviços nativos do Django: auth, sessions, content-type, CSRF-cookie
+- Implementação de Middleware personalizado para autenticação de tokens
+- Implementação de cache com memoria local -> futuramente implementar Redis